login y roles y permisos ajustados

+++ Odin/src/main/java/com/odin/odin/controller/EstadosController.java
@@ -5,43 +5,42 @@ import com.odin.odin.repository.EstadosRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-
 import java.util.List;
 
 @RestController
 @RequestMapping("/api/estados")
-public class EstadosController
-{
+public class EstadosController {
+
     @Autowired
     private EstadosRepository estadosRepository;
 
     @GetMapping
-    public List<Estados> getAll()
-    {
+    public List<Estados> getAll() {
         return estadosRepository.findAll();
     }
 
     @GetMapping("/{id}")
-    public Estados getById(@PathVariable Long id)
-    {
+    public Estados getById(@PathVariable Integer id) {
         return estadosRepository.findById(id).orElse(null);
     }
+
     @PostMapping
-    public Estados create(@RequestBody Estados estados)
-    {
+    public Estados create(@RequestBody Estados estados) {
         return estadosRepository.save(estados);
     }
 
     @PutMapping("/{id}")
-    public Estados update(@PathVariable long id, @RequestBody Estados estados)
-    {
+    public Estados update(
+            @PathVariable Integer id,
+            @RequestBody Estados estados) {
+
         estados.setId_estado(id);
+
         return estadosRepository.save(estados);
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable long id)
-    {
+    public void delete(@PathVariable Integer id) {
         estadosRepository.deleteById(id);
     }
 }
