+++ Odin/src/main/java/com/odin/odin/model/Tramites.java
@@ -21,25 +21,34 @@ public class Tramites {
     private Long idTramite;
 
     @NotBlank(message = "El nombre es obligatorio")
-    @Column(name = "nombre", nullable = false, length = 200)
+    @Column(
+            name = "nombre",
+            nullable = false,
+            length = 200
+    )
     private String nombre;
 
     @NotBlank(message = "La descripción es obligatoria")
-    @Column(name = "descripcion", nullable = false, columnDefinition = "text")
+    @Column(
+            name = "descripcion",
+            nullable = false,
+            columnDefinition = "text"
+    )
     private String descripcion;
 
-    // ✅ CORREGIDO: @Column explícito
     @Column(name = "id_dependencia_responsable")
     private Long idDependenciaResponsable;
 
-    // ✅ CORREGIDO: Solo el ID, no la relación completa
     @Column(name = "id_estado_inicial")
-    private Long idEstadoInicial;
+    private Integer idEstadoInicial;
 
     @Column(name = "dias_respuesta")
     private Integer diasRespuesta;
 
-    @Column(name = "prioridad_default", length = 20)
+    @Column(
+            name = "prioridad_default",
+            length = 20
+    )
     private String prioridadDefault;
 
     @Column(name = "requiere_respuesta")
@@ -58,22 +67,27 @@ public class Tramites {
     @Column(name = "fecha_limite")
     private LocalDateTime fechaLimite;
 
-    // ✅ Relación con Dependencias (solo lectura)
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "id_dependencia_responsable", insertable = false, updatable = false)
+    @JoinColumn(
+            name = "id_dependencia_responsable",
+            insertable = false,
+            updatable = false
+    )
     private Dependencias dependenciaResponsable;
 
-    // ✅ Relación con Estados (solo lectura)
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "id_estado_inicial", insertable = false, updatable = false)
+    @JoinColumn(
+            name = "id_estado_inicial",
+            insertable = false,
+            updatable = false
+    )
     private Estados estadoInicial;
 
-    // Métodos helper para mantener compatibilidad con código existente
-    public Long getIdEstadoInicial() {
+    public Integer getIdEstadoInicial() {
         return idEstadoInicial;
     }
 
-    public void setIdEstadoInicial(Long idEstadoInicial) {
+    public void setIdEstadoInicial(Integer idEstadoInicial) {
         this.idEstadoInicial = idEstadoInicial;
     }
 
@@ -81,7 +95,10 @@ public class Tramites {
         return idDependenciaResponsable;
     }
 
-    public void setIdDependenciaResponsable(Long idDependenciaResponsable) {
-        this.idDependenciaResponsable = idDependenciaResponsable;
+    public void setIdDependenciaResponsable(
+            Long idDependenciaResponsable) {
+
+        this.idDependenciaResponsable =
+                idDependenciaResponsable;
     }
 }
